ai/requirements.txt
flask
flask-cors
werkzeug
flask-bcrypt
flask-jwt-extended
psycopg2-binary
requests
transformers
torch
torchvision
pillow
tensorflow-cpu
mangum